+++ beestation.dme
@@ -3741,7 +3741,7 @@
 #include "code\modules\modular_computers\file_system\programs\budgetordering.dm"
 #include "code\modules\modular_computers\file_system\programs\card.dm"
 #include "code\modules\modular_computers\file_system\programs\cargobounty.dm"
-#include "code\modules\modular_computers\file_system\programs\cdrmonitor.dm"
+#include "code\modules\modular_computers\file_system\programs\cdr_monitor.dm"
 #include "code\modules\modular_computers\file_system\programs\configurator.dm"
 #include "code\modules\modular_computers\file_system\programs\crewmanifest.dm"
 #include "code\modules\modular_computers\file_system\programs\file_browser.dm"
@@ -3898,8 +3898,8 @@
 #include "code\modules\power\apc\apc_mapping.dm"
 #include "code\modules\power\apc\apc_power_proc.dm"
 #include "code\modules\power\apc\apc_tool_act.dm"
+#include "code\modules\power\condensate_decay_reactor\condensate_decay_reactor.dm"
 #include "code\modules\power\condensate_decay_reactor\condensate_gas.dm"
-#include "code\modules\power\condensate_decay_reactor\condensatedecayreactor.dm"
 #include "code\modules\power\lighting\_light_defines.dm"
 #include "code\modules\power\lighting\light.dm"
 #include "code\modules\power\lighting\light_construct.dm"
