+++ code/modules/power/condensate_decay_reactor/condensate_decay_reactor.dm
