+++ code/modules/power/singularity/singularity.dm
@@ -34,7 +34,7 @@
 
 CREATION_TEST_IGNORE_SUBTYPES(/obj/anomaly/singularity)
 
-/obj/anomaly/singularity/Initialize(mapload, starting_energy = 100)
+/obj/anomaly/singularity/Initialize(mapload, starting_energy = 50)
 	. = ..()
 	START_PROCESSING(SSsinguloprocess, src)
 	AddElement(/datum/element/point_of_interest)
@@ -447,7 +447,7 @@ CREATION_TEST_IGNORE_SUBTYPES(/obj/anomaly/singularity/deadchat_controlled)
 	var/time_to_explode = 30 SECONDS
 
 /obj/anomaly/singularity/temporary/Initialize(mapload, starting_energy = 50)
-	. = ..(mapload, starting_energy)
+	. = ..()
 	addtimer(CALLBACK(src, PROC_REF(explode)), time_to_explode)
 
 /obj/anomaly/singularity/temporary/proc/explode()
